+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Dropped support for node v18 (\#930).
 * Added support for guest user (\#913).
 * Run some GitHub Actions when targeting `ihb-develop` branch (\#922).
+* Remove support for `Task.source` (\#931).
 
 # 1.22.2